% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{my_geom_defaults}
\alias{my_geom_defaults}
\alias{set_my_geom_defaults}
\title{My geoms}
\usage{
my_geom_defaults(
  ink_colour = "#333333",
  paper_colour = NA,
  shadow_colour = "#F2F2F2",
  font = "Roboto",
  small_text = 9,
  big_text = 12
)

set_my_geom_defaults(
  ink_colour = "#333333",
  paper_colour = NA,
  shadow_colour = "#F2F2F2",
  font = "Roboto",
  small_text = 9,
  big_text = 12
)
}
\arguments{
\item{ink_colour, paper_colour, shadow_colour}{A \code{character(1)} for colours.}

\item{font}{A \code{character(1)} with a font family name.}

\item{small_text, big_text}{A \code{numeric(1)} indicating font sizes.}
}
\value{
For \code{my_geom_defaults}: a \code{list} with defaults. For
\code{set_my_geom_defaults()}: nothing but global geom defaults will have
changed.
}
\description{
These are functions for getting and setting geom defaults.
}
\examples{
# Standard plot
p <- ggplot(mtcars, aes(mpg, disp)) +
  geom_point()

# Change default foreground colour to red
set_my_geom_defaults(ink_colour = "red")
p
}
