% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{my_theme}
\alias{my_theme}
\alias{set_my_theme}
\title{My theme}
\usage{
my_theme(
  ink_colour = "#333333",
  paper_colour = NA,
  shadow_colour = "#F2F2F2",
  font = "Roboto",
  small_text = 9,
  big_text = 12
)

set_my_theme(
  ink_colour = "#333333",
  paper_colour = NA,
  shadow_colour = "#F2F2F2",
  font = "Roboto",
  small_text = 9,
  big_text = 12
)
}
\arguments{
\item{ink_colour, paper_colour, shadow_colour}{A \code{character(1)} for colours.}

\item{font}{A \code{character(1)} with a font family name.}

\item{small_text, big_text}{A \code{numeric(1)} indicating font sizes.}
}
\value{
For \code{my_theme()}: A \code{theme} object. For \code{set_my_theme()}, nothing but
global theme settings will have changed.
}
\description{
These are functions for getting and setting my theme.
}
\examples{
# Standard plot
p <- ggplot(mtcars, aes(mpg, disp)) +
  geom_point(aes(colour = factor(cyl)))

# Applying theme
p + my_theme()

# Alternatively, to set globally:
set_my_theme()
p
}
