% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplotdefaults-package.R
\docType{package}
\name{myplotdefaults-package}
\alias{myplotdefaults}
\alias{myplotdefaults-package}
\title{myplotdefaults: Set plot defaults}
\description{
This is simply a repo to keep track of my plot defaults instead of having to declare it at the beginning of every script.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{tahvdbrand@gmail.com} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
