% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_defaults}
\alias{set_defaults}
\title{Set plot defaults}
\usage{
set_defaults(
  ink_colour = "#333333",
  paper_colour = NA,
  shadow_colour = "#F2F2F2",
  font = "Roboto",
  small_text = 9,
  big_text = 12
)
}
\arguments{
\item{ink_colour, paper_colour, shadow_colour}{A \code{character(1)} for colours.}

\item{font}{A \code{character(1)} with a font family name.}

\item{small_text, big_text}{A \code{numeric(1)} indicating font sizes.}
}
\value{
Nothing, but sets plot defaults
}
\description{
This function sets plot defaults, both geom defaults and theme defaults.
}
\examples{
NULL
}
